+++ CHANGELOG.md
@@ -9,23 +9,37 @@
 * Add a README.md
 
 ## 0.0.3
+
 * Add multiple ways to listen for loading paywalls (the AppStore ones)
 
-
 ## 0.0.4
+
 * Add initial event to paywallsDidLoad stream, so that if you start listening to it after the paywalls are loaded, you'll get the event
 
 ## 0.0.5
+
 * Change stream to broadcast, improve its lifecycle
 
 ## 0.0.5-1
+
 * Fix a bug where second stream listener would not get initial event
 
 ## 0.0.6
+
 * hasActiveSubscription method
 
 ## 0.0.7
+
 * update Apphud SDK
 
 ## 0.0.8
+
 * Change versioning to '~>' operator for automatic updates
+
+## 0.0.9
+
+* Add SKError to PurchaseResult
+
+## 0.0.10
+
+* Fix parsing error